+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+- Use `require` instead of `import` in patches (`import` doesn't work in Bun) (closes #205) (#200) - @bl-ue
+
 ## [v3.1.0](https://github.com/Piebald-AI/tweakcc/releases/tag/v3.1.0) - 2025-11-15
 
 - Show the currently-active toolset next to the "accept edits on"/"plan mode on" banner (#200) - @bl-ue